requirements.txt
Django==5.0.7
mysqlclient==2.2.4
python-decouple==3.8
Pillow==10.4.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-extensions==3.2.3
requests==2.32.3
celery==5.3.1
redis==4.6.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
whitenoise==6.5.0
gunicorn==21.2.0
